Fix main module - separate server from application.
