+++ package.json
@@ -2,7 +2,7 @@
   "name": "climbingcompetition",
   "version": "1.0.0",
   "description": "Climbing competition web application created with node.js",
-  "main": "src/server.js",
+  "main": "src",
   "scripts": {
     "start": "node %npm_package_main%",
     "test": "nyc mocha src/**/*.spec.js test/*.js",
@@ -51,7 +51,6 @@
   "devDependencies": {
     "codecov": "^3.0.0",
     "expect.js": "^0.3.1",
-    "js-mock": "^1.0.1",
     "mocha": "^5.0.5",
     "nyc": "^11.6.0",
     "supertest": "^3.0.0"
